Senorita/ytmplayer/requirements.txt
PySide6
ytmusicapi
python-mpv
requests
Pillow
yt-dlp
